py_rpc/requirements.txt
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.23.3
pydantic==1.10.9
python-dotenv==1.0.0
typing_extensions==4.6.3
